requirements.txt
# Core dependencies
openai>=1.12.0
requests>=2.31.0
pyyaml>=6.0.1

# Framework-specific dependencies
crewai>=0.1.0
langgraph>=0.0.1
langchain>=0.1.0
langchain-openai>=0.0.5
dspy-ai>=2.4.0

# Utilities
python-dotenv>=1.0.0
pandas>=2.0.0
tabulate>=0.9.0

